refactor(runtime): using macro to build metrics (#85)
* refactor(metrics): using seperate module for runtime metrics


* feat(protocol): add method to get endpoint


* feat(proxy): add name to listener


* refactor(runtime): using macro to build metrics
